+++ modules/carto/src/sources/h3-table-source.ts
@@ -1,6 +1,12 @@
 /* eslint-disable camelcase */
 import {baseSource} from './base-source';
-import type {AggregationOptions, SourceOptions, TableSourceOptions, TilejsonSource} from './types';
+import type {
+  AggregationOptions,
+  SourceOptions,
+  SpatialDataType,
+  TableSourceOptions,
+  TilejsonResult
+} from './types';
 
 export type H3TableSourceOptions = SourceOptions & TableSourceOptions & AggregationOptions;
 
@@ -8,21 +14,27 @@ type UrlParameters = {
   aggregationExp: string;
   aggregationResLevel?: string;
   columns?: string;
-  geo_column?: string;
+  spatialDataType: SpatialDataType;
+  spatialDataColumn?: string;
   name: string;
 };
 
-const h3TableSource: TilejsonSource<H3TableSourceOptions> = async function (
+export const h3TableSource = async function (
   options: H3TableSourceOptions
-): Promise<any> {
+): Promise<TilejsonResult> {
   const {
     aggregationExp,
     aggregationResLevel = 4,
     columns,
-    spatialDataColumn = 'h3:h3',
+    spatialDataColumn = 'h3',
     tableName
   } = options;
-  const urlParameters: UrlParameters = {aggregationExp, name: tableName};
+  const urlParameters: UrlParameters = {
+    aggregationExp,
+    name: tableName,
+    spatialDataColumn,
+    spatialDataType: 'h3'
+  };
 
   if (aggregationResLevel) {
     urlParameters.aggregationResLevel = String(aggregationResLevel);
@@ -30,10 +42,5 @@ const h3TableSource: TilejsonSource<H3TableSourceOptions> = async function (
   if (columns) {
     urlParameters.columns = columns.join(',');
   }
-  if (spatialDataColumn) {
-    urlParameters.geo_column = spatialDataColumn;
-  }
-  return baseSource<UrlParameters>('table', options, urlParameters);
+  return baseSource<UrlParameters>('table', options, urlParameters) as Promise<TilejsonResult>;
 };
-
-export {h3TableSource};