+++ CHANGELOG.md
@@ -21,6 +21,14 @@ Ref: http://keepachangelog.com/en/0.3.0/
 
 ### deck.gl v8.10 Prereleases
 
+#### deck.gl [8.10.0-alpha.6] - Oct 18 2023
+
+ - Fix stylesheet.css (#8210)
+
+#### deck.gl [8.10.0-alpha.5] - Oct 17 2023
+
+ - [chore] Fix widget module publishing (#8178)
+
 #### deck.gl [8.10.0-alpha.4] - Oct 5 2023
 
 - add(widgets) fullscreen widget (#8024)